+++ crates/mirror-mirror/src/foreign_impls/btree_map.rs
@@ -4,6 +4,7 @@ use core::any::Any;
 use core::fmt;
 
 use crate::iter::PairIterMut;
+use crate::map::MapError;
 use crate::type_info::graph::MapNode;
 use crate::type_info::graph::NodeId;
 use crate::type_info::graph::TypeGraph;
@@ -21,52 +22,7 @@ where
     K: FromReflect + DescribeType + Ord,
     V: FromReflect + DescribeType,
 {
-    fn get(&self, key: &dyn Reflect) -> Option<&dyn Reflect> {
-        let key = K::from_reflect(key)?;
-        let value = self.get(&key)?;
-        Some(value.as_reflect())
-    }
-
-    fn get_mut(&mut self, key: &dyn Reflect) -> Option<&mut dyn Reflect> {
-        let key = K::from_reflect(key)?;
-        let value = self.get_mut(&key)?;
-        Some(value.as_reflect_mut())
-    }
-
-    fn insert(&mut self, key: &dyn Reflect, value: &dyn Reflect) -> Option<Box<dyn Reflect>> {
-        let key = K::from_reflect(key)?;
-        let value = V::from_reflect(value)?;
-        let previous = BTreeMap::insert(self, key, value)?;
-        Some(Box::new(previous))
-    }
-
-    fn remove(&mut self, key: &dyn Reflect) -> Option<Box<dyn Reflect>> {
-        let key = K::from_reflect(key)?;
-        let previous = BTreeMap::remove(self, &key)?;
-        Some(Box::new(previous))
-    }
-
-    fn len(&self) -> usize {
-        self.len()
-    }
-
-    fn is_empty(&self) -> bool {
-        self.is_empty()
-    }
-
-    fn iter(&self) -> crate::map::Iter<'_> {
-        let iter = self
-            .iter()
-            .map(|(key, value)| (key.as_reflect(), value.as_reflect()));
-        Box::new(iter)
-    }
-
-    fn iter_mut(&mut self) -> PairIterMut<'_, dyn Reflect> {
-        let iter = self
-            .iter_mut()
-            .map(|(key, value)| (key.as_reflect(), value.as_reflect_mut()));
-        Box::new(iter)
-    }
+    map_methods!();
 }
 
 impl<K, V> DescribeType for BTreeMap<K, V>